+++ lib/gooddata/lcm/lcm2.rb
@@ -71,6 +71,7 @@ module GoodData
         SynchronizeMeta,
         SynchronizeProcesses,
         SynchronizeSchedules,
+        SynchronizeColorPalette,
         SynchronizeNewSegments,
         UpdateReleaseTable
       ],