+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support for Automated Data Distribution in project
 - Added Dockefile for easy build of dockerized gooddata-ruby
 - Handle export/import async task errors (TMA-231)
+- Support for custom color palette
 
 ## 0.6.49
 - Implement Helpers::GD_MAX_RETRY to allow max retries override